connect_memory_with_llm.py
import os

from langchain_huggingface import HuggingFaceEndpoint
from langchain_core.prompts import PromptTemplate
from langchain.chains import RetrievalQA
from langchain_huggingface import HuggingFaceEmbeddings
from langchain_community.vectorstores import FAISS
from dotenv import load_dotenv
load_dotenv()


# Step 1: Setup LLM (Mistral with HuggingFace)
HF_TOKEN=os.environ.get("HF_TOKEN")
HUGGINGFACE_REPO_ID="mistralai/Mistral-7B-Instruct-v0.3"

def load_llm(huggingface_repo_id):
    llm=HuggingFaceEndpoint(
        repo_id=huggingface_repo_id,
        temperature=0.5,
        model_kwargs={"token":HF_TOKEN,
                      "max_length":"512"}
    )
    return llm

# Step 2: Connect LLM with FAISS and Create chain

CUSTOM_PROMPT_TEMPLATE = """You are given a set of contextual information that serves as the sole reference for answering the user's question. Your task is to generate a detailed and accurate response using only the information provided in the context, ensuring that the answer remains strictly within its scope.

Guidelines for Answer Generation:
Start the answer directly without introductory phrases or small talk.
Provide a thorough explanation while maintaining relevance—ensure the response is detailed but does not introduce any information that is not explicitly present in the context.
If the context does not contain sufficient details to answer the question, explicitly state that you do not know the answer rather than making assumptions or extrapolations.
Ensure clarity, coherence, and completeness in the response while keeping it fully aligned with the question.
Context: {context}
Question: {question}

Generate a well-structured and contextually accurate response accordingly."""

def set_custom_prompt(custom_prompt_template):
    prompt=PromptTemplate(template=custom_prompt_template, input_variables=["context", "question"])
    return prompt

# Load Database
DB_FAISS_PATH="vectorstore/db_faiss"
embedding_model=HuggingFaceEmbeddings(model_name="sentence-transformers/all-MiniLM-L6-v2")
db=FAISS.load_local(DB_FAISS_PATH, embedding_model, allow_dangerous_deserialization=True)

# Create QA chain
qa_chain=RetrievalQA.from_chain_type(
    llm=load_llm(HUGGINGFACE_REPO_ID),
    chain_type="stuff",
    retriever=db.as_retriever(search_kwargs={'k':3}),
    return_source_documents=True,
    chain_type_kwargs={'prompt':set_custom_prompt(CUSTOM_PROMPT_TEMPLATE)}
)

# Now invoke with a single query
user_query=input("Write Query Here: ")
response=qa_chain.invoke({'query': user_query})
print("\n\nRESULT: ", response["result"])
print("\n\nSOURCE DOCUMENTS: ", response["source_documents"])