requirements.txt
python-dotenv
langchain
langchain_community
langchain_huggingface
faiss-cpu
pypdf
streamlit